refactor(userspace-dp): relocate frame/wg.rs inline tests to sibling file
frame/wg.rs was 1561 LOC: production WG-encap code in lines ~1-603 and a
single inline `#[cfg(test)] mod wg_frame_tests { ... }` block (lines
604-1561) carrying all 19 `#[test]` functions. Per codex-review-174 #8
the production WG encap is DELIBERATELY not split — splitting it risks
hiding the single-underlay-FIB-lookup invariant (#3992) and inviting
duplicate lookups — so only the test module is extracted.

Move the inline test module body verbatim (a 4-space dedent out of the
`mod {}` wrapper, no test logic changed) to a sibling `frame/wg_tests.rs`
and replace the inline block with `#[cfg(test)] #[path = "wg_tests.rs"]
mod tests;`, the same convention as #4668 (reject_reply) and #4670
(dispatch).

The two top-level `#[cfg(test)]` helpers that live OUTSIDE the mod block
stay in wg.rs: `OUTER_ROUTE_RESOLVE_COUNT` (the #3992 resolve-once canary
counter, `pub(super)`, incremented by the production egress helper) and
`udp6_checksum_scalar_reference` (the parity reference oracle). The
relocated tests reach both via `use super::*`.

Production lines 1-603 are byte-identical (git diff: 2 insertions + 957
deletions = mod block removed, 3-line declaration added). Relocated test
bodies are byte-identical after reversing the dedent. Test count is
19 before == 19 after (now under `afxdp::frame::wg::tests::`).

Validation: cargo build ok; full suite `cargo test --release
--test-threads=1` reports 3860 passed / 0 failed / 2 ignored (matches
the pre-split baseline); the #3992 canary
`wg_encap_frame_resolves_outer_route_once_v4` passes.

Closes #4671
